+++ src/Connector.jsx
@@ -57,19 +57,37 @@ export const Connector = () => {
     setRunning(false);
   };
 
-  const reset = async () => {
-    await esploader().hard_reset();
+  const erase = async () => {
+    try {
+      setRunning(true);
+      await new Promise((resolve) => setTimeout(resolve, 3000));
+      await esploader().erase_flash();
+    } catch (e) {
+      console.error(e);
+      term.writeln(`Error: ${e.message}`);
+    } finally {
+      setRunning(false);
+    }
   };
 
   return (
     <div id="connector">
+      <h3>Device Connection</h3>
       <Show when={!connected()}>
-        <button disabled={running()} onClick={connect}>Connect</button>
+        <button disabled={running()} onClick={connect}>
+          Connect to Device
+        </button>
       </Show>
       <Show when={connected()}>
-        <button disabled={running()} onClick={disconnect}>Disconnect</button>
-        <button disabled={running()} onClick={reconnect}>Reconnect</button>
-        <button disabled={running()} onClick={reset}>Reset</button>
+        <button disabled={running()} onClick={disconnect}>
+          Disconnect from Device
+        </button>
+        <button disabled={running()} onClick={reconnect}>
+          Reconnect
+        </button>
+        <button disabled={running()} onClick={erase}>
+          Erase Firmware
+        </button>
       </Show>
     </div>
   );
